Migration of original scripts to JavaScript
These files are the same as the original python files on the master branch of this repository except that they are in JavaScript and now output additional information to the databases that assist in integrating them with Active Documentation. Additionally, each .java can be processed as a separate XML files.