Changes to databases info format
The databases are no longer output to files. Instead, their information is placed in an array in which the first entry is the name of the database (the file name), and the second entry is the information that used to be output to the file (i.e. the database itself).